requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openpyxl
pillow
pytesseract
jinja2
